% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_idx.R
\name{add_event_idx}
\alias{add_event_idx}
\title{Add event index}
\usage{
add_event_idx(df, step2time)
}
\arguments{
\item{df}{a dataframe with at least 2 columns for date and time. There should be no replication in date and time across rows.}

\item{step2time}{the sampling frequency of the dataframe.}
}
\value{
the input dataframe with an added 'event' column with the event id.
}
\description{
Adds a column to the input dataframe with event id number according to the continuity of date and time columns.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
