% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities.R
\name{add_velocities_parallel}
\alias{add_velocities_parallel}
\title{Velocity calculations in parallel}
\usage{
add_velocities_parallel(
  data,
  pos_label_x = "lon",
  pos_label_y = "lat",
  id_label = "id",
  sample_step = 1,
  min_displacement = 0.1,
  lonlat = FALSE,
  verbose = FALSE,
  step2time = 1
)
}
\arguments{
\item{data}{Time series with individual's positional data through time. Columns must include: id, time, lon, lat.}

\item{pos_label_x}{column name of x position column, default = 'lon'.}

\item{pos_label_y}{column name of y position column, default = 'lat'.}

\item{id_label}{column name of individual local identifier column, default = 'id'.}

\item{sample_step}{An integer, the step over which to calculate the velocities, in timesteps.}

\item{min_displacement}{Minimum distance traveled to constitute movement, default = '0.1'.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}

\item{step2time}{the sampling frequency, the relation between a time step and real time in seconds}
}
\value{
a list of velocities over time
}
\description{
Calculates velocity and headings based on two location points and the time taken to travel between those points
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
