% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vels.R
\name{add_vels}
\alias{add_vels}
\title{Add motion properties in parallel}
\usage{
add_vels(data, lonlat = TRUE, verbose = FALSE, parallelize = FALSE)
}
\arguments{
\item{data}{Time series with individual's positional data through time.
Columns must include: id, time, lon, lat.}

\item{lonlat}{whether positions are geographic coordinates, default = TRUE.}

\item{verbose}{whether to post updates on progress, default = FALSE.}

\item{parallelize}{whether to run the function in parallel, default = FALSE.}
}
\value{
the input dataframe with new speed and heading (rotational) column
}
\description{
Calculates headings and speeds based on
two location points and the time taken to travel between those points
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
