% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_shape.R
\name{bb_angl_dev}
\alias{bb_angl_dev}
\title{Bounding box angle deviation}
\usage{
bb_angl_dev(flock_vel, oobb)
}
\arguments{
\item{flock_vel}{a 2D vector representing the group's velocity at one timestep}

\item{oobb}{the object oriented bounded box from the positions of all group members}
}
\description{
Get the minimum deviation of the group's velocity and the bounding box of individual positions
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
