% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing_angle.R
\name{bearing_angle}
\alias{bearing_angle}
\title{Bearing angle}
\usage{
bearing_angle(h_a, pos_a, pos_b)
}
\arguments{
\item{h_a}{A 2D vector representing the heading of focal individual.}

\item{pos_a}{A 2D vector representing the position of the focal individual.}

\item{pos_b}{A 2D vector representing the position of the neighboring individual.}
}
\value{
a double, the bearing angle from individual a to individual b
}
\description{
Calculates the bearing angle from a focal individual (a) to a neighbor (b).
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
