% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_metrics.R
\name{calc_events_averages}
\alias{calc_events_averages}
\title{Calculate events averages}
\usage{
calc_events_averages(df)
}
\arguments{
\item{df}{data frame with time series of group measurements
Columns must include:  event, set, group_size, mean_nnd,
sd_nnd, pol, speed, sd_front, mean_bangl, and shape.}
}
\value{
a dataframe with a metric value per event.
}
\description{
Calculating all metrics per event
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
