% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{calc_events_duration}
\alias{calc_events_duration}
\title{Events total duration}
\usage{
calc_events_duration(data, step2time)
}
\arguments{
\item{data}{A dataframe with a keep (representing which rows are defined as events of collective motion) and a set column}

\item{step2time}{the sampling frequency of the dataframe.}
}
\value{
a double, the total duration of events in the dataset
}
\description{
Calculates the total duration of events of collective motion in a dataset
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
