% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{calc_events_N}
\alias{calc_events_N}
\title{Number of events}
\usage{
calc_events_N(data)
}
\arguments{
\item{data}{A dataframe with a keep (representing which rows are defined as events of collective motion) and a date column}
}
\value{
an integer with the number of events of colletive motion
}
\description{
Calculates the number of events of collective motion in dataset
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
