% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontness.R
\name{calc_frontness}
\alias{calc_frontness}
\title{Frontness}
\usage{
calc_frontness(bangls)
}
\arguments{
\item{bangls}{A vector of bearing angles in rads.}
}
\value{
a vector with doubles, the frontness metric (values 0 to 1) relative to the input vector
}
\description{
Calculates how in front or behind the neighbor is based on the bearing angle from the focal individual to the neighbor.
}
\author{
Marina Papadopoulou

Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
