% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_metrics_calc.R
\name{calc_global_metrics}
\alias{calc_global_metrics}
\title{Calculation of group metrics}
\usage{
calc_global_metrics(x, lonlat)
}
\arguments{
\item{x}{A dataframe timestep of a group.}

\item{lonlat}{logical, whether positions are
geographic coordinates, default = FALSE.}
}
\value{
A dataframe with the group average timeseries, with columns:
 set, t, pol, speed, shape, N (number of individuals),
 missing_ind (whether some individuals are missing).
}
\description{
Calculates the average speed,
 polarization and shape of the group through time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
