% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_metrics_per_event.R
\name{calc_metrics_per_event}
\alias{calc_metrics_per_event}
\title{Calculate metrics of collective motion}
\usage{
calc_metrics_per_event(global_df, pairwise_df)
}
\arguments{
\item{global_df}{A data frame with time series of global group measurements. Columns must include:  date, time, event, pol, shape, speed}

\item{pairwise_df}{A data frame with time series of pairwise measurements. Columns must include:  date, time, id, dist, bangl}
}
\value{
A dataframe with 10 metrics per event
}
\description{
Calculates metrics of collective motion across dates and events
}
\seealso{
\code{\link{group_motion_timeseries}}, \code{\link{group_metrics_parallel}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
