% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{calc_n_events}
\alias{calc_n_events}
\title{Number of Events (Internal)}
\usage{
calc_n_events(data)
}
\arguments{
\item{data}{The dataframe of one set (splitted by the \code{\link{events_n}}
function) with a \code{keep} column.}
}
\value{
An integer corresponding to the number of events of collective motion
 (sequences of \code{keep == TRUE}).
}
\description{
This function calculates the number of events of collective
 motion per set, given by \code{\link{events_n}}.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
