% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swaRm_group_shape.R
\name{calc_obb}
\alias{calc_obb}
\title{Minimum Bounding Box}
\usage{
calc_obb(x, y, geo = FALSE)
}
\arguments{
\item{x}{A vector of x (or longitude) coordinates.}

\item{y}{A vector of y (or latitude) coordinates.}

\item{geo}{A logical value indicating whether the locations are defined by
geographic coordinates (pairs of longitude/latitude values). Default: FALSE.}
}
\value{
A list with the bounding box coordinates, its heights, its width, 
and the orientation of its longest side in degrees.
}
\description{
Calculates the minimum bounding box of a group of objects
}
\seealso{
\code{\link{is_chull}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}, 
arina Papadopoulou, \email{m.papadopoulou.rug@gmail.com}
}
