% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_metrics.R
\name{calc_pairwise_metrics}
\alias{calc_pairwise_metrics}
\title{Pairwise Metrics}
\usage{
calc_pairwise_metrics(df)
}
\arguments{
\item{df}{A data frame with time series of pairwise measurements. Columns
must include: \code{set}, \code{t}, \code{id}, \code{nnd}, \code{bangl}.}
}
\value{
A dataframe with average and standard deviation of pairwise
 measurements for a group at a single time step.
}
\description{
This function calculates group averages of pairwise measures.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
