% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_col_motion_metrics.R
\name{col_motion_metrics_from_raw}
\alias{col_motion_metrics_from_raw}
\title{Collective motion metrics from raw data}
\usage{
col_motion_metrics_from_raw(
  data,
  mov_av_time_window,
  speed_sliding_window = 1,
  step2time = 1,
  lonlat = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame with time series of individual's positional data through time. Columns must include: id, date, time, posx, posy}

\item{mov_av_time_window}{to average over (in timesteps)}

\item{speed_sliding_window}{An integer, the step over which to calculate the velocities, in timesteps.}

\item{step2time}{the sampling frequency, the relation between a time step and real time in seconds}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
either a list of dataframes with neighbor ids, bearing angles, distances and heading deviations for each individual through time, or saves individual csvs per day, depending on input.
}
\description{
Calculates metrics of collective motion across dates and events
}
\seealso{
\code{\link{neighb_rel_pos_timeseries_parallel}}, \code{\link{group_motion_timeseries}}, \code{\link{group_metrics_parallel}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
