% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_motion_timeseries.R
\name{do_add_motion_properties}
\alias{do_add_motion_properties}
\title{Adding motion properties in parallel without progress updates}
\usage{
do_add_motion_properties(data, lonlat = FALSE)
}
\arguments{
\item{data}{Time series with individual's positional data through time. Columns must include: id, time, lon, lat.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}
}
\value{
the input dataframe with new speed and heading (rotational) column
}
\description{
Calculates headings and speeds based on two location points and the time taken to travel between those points
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
