% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_swarm_space.R
\name{do_tsne}
\alias{do_tsne}
\title{Performs t-SNE}
\usage{
do_tsne(df, tsne_rand_seed, tsne_perplexity)
}
\arguments{
\item{df}{A dataframe with the metrics of collective motion
per event.}

\item{tsne_rand_seed}{Random seed to be able to reproduce the t-SNE analysis.}

\item{tsne_perplexity}{Perplexity parameter for tsne. Usually between 10-50.}
}
\value{
A dataframe with X and Y coordinates per event in the t-SNE space.
}
\description{
Calculating the coordinates of each event in the t-SNE space.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
