% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_idx.R
\name{event_ids}
\alias{event_ids}
\title{Event indexies}
\usage{
event_ids(t, step2time)
}
\arguments{
\item{t}{a vector of time.}

\item{step2time}{the sampling frequency of t.}
}
\value{
a vector of the same size as t with event ids
}
\description{
Returns a vector with event ids that corresponds to the input time vector, according to the continuity of time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
