% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_idx.R
\name{event_ids_per_set}
\alias{event_ids_per_set}
\title{Event indexies per set}
\usage{
event_ids_per_set(setdf, ev_idx)
}
\arguments{
\item{t}{a vector of time.}
}
\value{
a vector of the same size as t with event ids
}
\description{
Returns a vector with event ids that corresponds
to the input time vector, according to the continuity of time.
}
\details{
#' @title Event indexies
#' @description Returns a vector with event ids that corresponds
#' to the input time vector, according to the continuity of time.
#' @param df A dataframe with a set and t columns
#' for adding event ids.
#' @param step2time the sampling frequency of t.
#' @return a vector of the same size as t with event ids
#' @author Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
#' @export
event_ids <- function(df, step2time) {
  counted_ev <- 0
  sets <- unique(df$set)
  event_idxs <- c()
  for (i in sets){
    counted_ev <- length(unique(event_idxs))
    ev_ids <- event_ids_per_set(df[df$set == i, "t"], step2time)
    ev_ids <- ev_ids + counted_ev
    event_idxs <- c(event_idxs, ev_ids)
  }
  return(event_idxs)
}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
