% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_metrics.R
\name{event_metrics}
\alias{event_metrics}
\title{Metrics of Collective Motion}
\usage{
event_metrics(global_df, pairwise_df)
}
\arguments{
\item{global_df}{A data frame with time series of global group measurements.
Columns must include: \code{set}, \code{t}, \code{event}, \code{pol},
\code{shape}, \code{speed}.}

\item{pairwise_df}{A data frame with time series of pairwise measurements.
Columns must include: \code{set}, \code{t}, \code{id}, \code{nnd},
\code{bangl}.}
}
\value{
A dataframe with 10 metrics per event.
}
\description{
This function calculates metrics of collective motion across
 sets and events.
}
\seealso{
\code{\link{group_metrics}}, \code{\link{nn_metrics}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
