% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{events_dur_summary}
\alias{events_dur_summary}
\title{Events duration summry}
\usage{
events_dur_summary(data, date_label, step2time)
}
\arguments{
\item{data}{A dataframe with a keep (representing which rows are defined as events of collective motion) and a date column}

\item{date_label}{the name of the column with date information}

\item{step2time}{the sampling frequency of the dataframe.}
}
\value{
a dataframe with 3 columns: date, ev_count (number of events), dur (duration of events in seconds)
}
\description{
Provides a summary with the number of events, their start time and duration in the dataset
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
