% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{events_n}
\alias{events_n}
\title{Number of events}
\usage{
events_n(data)
}
\arguments{
\item{data}{A dataframe with a keep column
(representing which rows are defined as events of collective motion)
and a set column.}
}
\value{
an integer with the number of events of collective motion
(sequences of keep == TRUE).
}
\description{
Calculates the number of events of collective motion
 in a dataset.
}
\seealso{
\code{\link{define_events}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
