% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_swarm_space.R
\name{expand_pca_swarm_space}
\alias{expand_pca_swarm_space}
\title{Expands existing swarm space (PCA)}
\usage{
expand_pca_swarm_space(
  metrics_data,
  pca_space,
  pca_data,
  event_duration_limit = NA
)
}
\arguments{
\item{metrics_data}{New metrics data to add in swarm space.}

\item{pca_space}{The PCA output to predict from.}

\item{pca_data}{Coordinates of initial data in swarm space.}

\item{event_duration_limit}{Filter out events that are shorter
than given duration.}
}
\value{
Τhe extended swarm space, x and y coordinates
per event of each species.
}
\description{
Predicts the positions of new event data in existing PCA space.
}
\seealso{
\code{\link{create_swarm_space}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
