% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events_summary.R
\name{get_events_summary}
\alias{get_events_summary}
\title{Get events summary}
\usage{
get_events_summary(df, step2time, species_id)
}
\arguments{
\item{df}{A dataframe, needs to have a keep and an event column}

\item{step2time}{the sampling frequency, the relation between a time step and real time in seconds}

\item{species_id}{A string representing the name of the species to add as a column}
}
\value{
a dataframe with event id, event duration in timesteps and in seconds, and the species name
}
\description{
Provides a summary of the number and duration of events of collective motion in a dataset.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
