% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_metrics_calc.R
\name{global_metrics}
\alias{global_metrics}
\title{Calculation of group metrics in parallel}
\usage{
global_metrics(data, lonlat, step2time = 1, parallelize = FALSE)
}
\arguments{
\item{data}{A dataframe of (ordered) time series of headings,
positions and speeds per individual. The dataframe may contain
several individuals. Should include column for:
id, time, speed, headx, heady, posx, posy.}

\item{lonlat}{logical, whether positions are geographic coordinates,
default = FALSE.}

\item{step2time}{the sampling frequency, the relation between a time
step and real time in seconds}

\item{parallelize}{whether to parallelize over time.}
}
\value{
A dataframe with the group average timeseries, with columns:
 set, time, pol, speed, shape, N (number of individuals),
 missing_ind (whether some individuals are missing).
}
\description{
Calculates the average speed, polarization
and shape of the group through time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
