% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_global_metrics.R
\name{group_global_metrics}
\alias{group_global_metrics}
\title{Global group metrics calculation}
\usage{
group_global_metrics(
  data_list,
  mov_av_time_window,
  geo,
  step2time,
  parallelize = FALSE
)
}
\arguments{
\item{data_list}{A list of dataframes with groups timeseries per day.
Column names must include: id, t, set, head, x, y.}

\item{mov_av_time_window}{Integer, timesteps to use as a sliding window
for average speed and polarization. Set NA if calculation is not needed.}

\item{geo}{logical, whether positions are geographic coordinates,
default = FALSE.}

\item{step2time}{the sampling frequency, the relation between a time
step and real time in seconds}

\item{parallelize}{whether or not to parallelize over ids.}
}
\value{
a dataframe with a column for neighbor id, bearing angle,
distance and heading deviation for each individual through time.
}
\description{
Calculates the average speed, polarization and
shape of all groups through time.
}
\seealso{
\code{\link{global_metrics}}, \code{\link{moving_average}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
