% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_nn_metrics.R
\name{group_nn_metrics}
\alias{group_nn_metrics}
\title{Nearest neighbor analysis}
\usage{
group_nn_metrics(
  data_list,
  lonlat = FALSE,
  verbose = FALSE,
  parallelize = FALSE
)
}
\arguments{
\item{data_list}{A list of dataframes with groups timeseries per day.
Column names must include: id, t, set, head, x, y.}

\item{lonlat}{logical, whether positions are geographic coordinates,
default = FALSE.}

\item{verbose}{whether to post updates on progress}

\item{parallelize}{whether to run the function in parallel over individuals,
default = FALSE.}
}
\value{
either a list of dataframes with neighbor ids, bearing angles,
distances and heading deviations for each individual through time.
}
\description{
Calculates the bearing angle and distance from a
focal individual of a group to its nearest neighbor over time.
}
\seealso{
\code{\link{nn_metrics}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
