% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarization.R
\name{group_polarization}
\alias{group_polarization}
\title{Group polarization}
\usage{
group_polarization(data)
}
\arguments{
\item{data}{A dataframe of (ordered) time series of heading per individual. The dataframe may contain several individuals.
Column names must be: id, headx, heady, time.}
}
\value{
A vector of polarization over time
}
\description{
Calculates the timeseries of polarization of a group
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
