% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_metrics.R
\name{group_props_calc}
\alias{group_props_calc}
\title{Calculation of group metrics}
\usage{
group_props_calc(x, lonlat)
}
\arguments{
\item{x}{A dataframe timestep of a group.}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}
}
\value{
A dataframe with the group average timeseries, with columns:
 date, time, pol, speed, shape, N (number of individuals), missing_ind (whether some individuals are missing).
}
\description{
Calculates the average speed, polarization and shape of the group through time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
