% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_shape.R
\name{group_shape_oblong_dev}
\alias{group_shape_oblong_dev}
\title{Group shape - Oblong deviation}
\usage{
group_shape_oblong_dev(df_self, pos_label_x = "posx", pos_label_y = "posy")
}
\arguments{
\item{df_self}{Dataframe with individual data over time, needs a time and position columns.}

\item{pos_label_x}{the name of the column with x coordinate of individual position}

\item{pos_label_y}{the name of the column with y coordinate of individual position}
}
\value{
a vector of doubles, the deviation (angle) of the bounding box of a group to its heading over time.
}
\description{
Calculates a metric of group shape: the deviation of the longest side of a bounding box around all group member from the groups' heading.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
