% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities_groups.R
\name{group_vels}
\alias{group_vels}
\title{Group velocity timeseries}
\usage{
group_vels(data, geo = TRUE, verbose = FALSE, parallelize = FALSE)
}
\arguments{
\item{data}{A data frame with time series of individual's positional data,
as exported by the \code{set_data_format} function.
Columns needed: 't', 'x', 'y', 'id', 'set'.}

\item{geo}{Logical, whether positions are geographic coordinates,
default = FALSE.}

\item{verbose}{Logical, whether to post updates on progress, default = FALSE.}

\item{parallelize}{Logical, whether to run the function in parallel over individuals,
default = FALSE.}
}
\value{
A list of dataframes, an element per set from the input dataframe
with new columns: head and speed.
}
\description{
Calculates and adds the speed and heading of each individual
over time in the dataset, and splits it in a list of dataframes based
on the defined sets.
}
\seealso{
\code{\link{add_vels}}, \code{\link{set_data_format}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
