% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_shape.R
\name{make_obb}
\alias{make_obb}
\title{Make oriented bounding box}
\usage{
make_obb(df_self_timestep, pos_label_x = "posx", pos_label_y = "posy")
}
\arguments{
\item{df_self_timestep}{Dataframe with focal individuals data for 1 timestep, needs position columns.}

\item{pos_label_x}{the name of the column with x coordinate of individual position}

\item{pos_label_y}{the name of the column with y coordinate of individual position}
}
\value{
a bounding box around all input positions
}
\description{
Creates the oriented bounded box of a group in one time step
}
\author{
Marina Papadopoulou

Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
