% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Moving average}
\usage{
moving_average(timeseries, window)
}
\arguments{
\item{timeseries}{Vector of doubles (timeseries)}

\item{window}{Time-window to average over (in timesteps)}
}
\value{
a vector of doubles (average over the window)
}
\description{
Calculates the moving average of a time series.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
