% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_species_metrics}
\alias{new_species_metrics}
\title{The Collective Motion Metrics of a New Species}
\format{
An object of class \code{data.frame} with 60 rows and 16 columns.
}
\usage{
data('new_species_metrics')
}
\description{
The output dataset of vignette 2, containing the metrics of 
 collective motion for a new species.
}
\keyword{datasets}
