% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_pairwise_analysis.R
\name{nn_analysis}
\alias{nn_analysis}
\title{Nearest neighbor analysis}
\usage{
nn_analysis(data_dates_list, lonlat = FALSE, verbose = FALSE)
}
\arguments{
\item{data_dates_list}{A list of dataframes with groups timeseries per day. Column names must include: id, time, date, headx, heady, posx, posy}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
either a list of dataframes with neighbor ids, bearing angles, distances and heading deviations for each individual through time.
}
\description{
Calculates the bearing angle and distance from a focal individual of a group to its nearest neighbor over time.
}
\seealso{
\code{\link{nn_rel_pos_timeseries_parallel}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
