% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_metrics_calc.R
\name{nn_metrics}
\alias{nn_metrics}
\title{Nearest neighbor metrics}
\usage{
nn_metrics(
  data,
  add_coords = FALSE,
  lonlat = FALSE,
  verbose = FALSE,
  parallelize = FALSE
)
}
\arguments{
\item{data}{Dataframe with group's timeseries for one set.
Column names must include: id, time.}

\item{add_coords}{whether data on relative positions are converted into
geographic coordinates, default = 'FALSE'.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}

\item{parallelize}{whether to parallelize the function over time}
}
\value{
a dataframe with a column for neighbor id, bearing angle,
distance and heading deviation for each individual through time.
}
\description{
Calculates the bearing angle and distance from
a focal individual of a group to its nearest neighbor over time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
