% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbour_pos.R
\name{nn_rel_pos_timeseries_parallel}
\alias{nn_rel_pos_timeseries_parallel}
\title{Relative position of nearest neighbor in parallel}
\usage{
nn_rel_pos_timeseries_parallel(
  data,
  add_coords = FALSE,
  lonlat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Dataframe with group's timeseries. Column names must include: id, time.}

\item{add_coords}{whether data is converted to geographic coordinates, default = 'FALSE'.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
a dataframe with a column for neighbor id, bearing angle, distance and heading deviation for each individual through time.
}
\description{
Calculates the bearing angle and distance from a focal individual of a group to its nearest neighbor over time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
