% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities.R
\name{padd_motion_properties}
\alias{padd_motion_properties}
\title{Add motion properties in parallel}
\usage{
padd_motion_properties(data, lonlat = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{Time series with individual's positional data through time. Columns must include: id, time, lon, lat.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
the input dataframe with new speed and heading (rotational) column
}
\description{
Calculates headings and speeds based on two location points and the time taken to travel between those points
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
