% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_pairwise_analysis.R
\name{pairwise_analysis}
\alias{pairwise_analysis}
\title{Pairwise analysis}
\usage{
pairwise_analysis(
  data_dates_list,
  max_n_rank_to_return = 4,
  return_df = TRUE,
  out_csv_dir = NA,
  add_coords = FALSE,
  lonlat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data_dates_list}{A list of dataframes with groups timeseries per day. Column names must include: id, time, date, headx, heady, posx, posy}

\item{max_n_rank_to_return}{Max rank of neighbor (e.g. 4th closest) to save/return data for}

\item{return_df}{logical, whether to return the result dataframes or not.}

\item{out_csv_dir}{directory output is saved to.}

\item{add_coords}{whether data is converted to geographic coordinates, default = 'FALSE'.}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
either a list of dataframes with neighbor ids, bearing angles, distances and heading deviations for each individual through time, or saves individual csvs per day, depending on input.
}
\description{
Calculates the bearing angle and distance from a focal individual of a group to all neighbors over time.
}
\seealso{
\code{\link{neighb_rel_pos_timeseries_parallel}}, \code{\link{bearing_angle}}, \code{\link{rad_between}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
