% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_pairwise.R
\name{pairwise_metrics}
\alias{pairwise_metrics}
\title{Pairwise metrics of collective motion in dataset}
\usage{
pairwise_metrics(
  data_list,
  geo = FALSE,
  verbose = FALSE,
  parallelize = FALSE,
  add_coords = FALSE
)
}
\arguments{
\item{data_list}{A list of dataframes with groups timeseries per set.
Columns must include: id, t, set, head, x, y.}

\item{geo}{Logical, whether positions are geographic coordinates,
default = FALSE.}

\item{verbose}{Logical, whether to post updates on progress, default = FALSE.}

\item{parallelize}{Logical, whether to run the function in parallel over
timesteps, default = FALSE.}

\item{add_coords}{Logical, whether data on relative positions are converted
into geographic coordinates, default = 'FALSE'.}
}
\value{
A dataframe format of the input list, with new columns for nearest neighbor id (nn_id),
bearing angles (bangl), and distances (nnd). If add_coords is TRUE, the columns
nnx and nny are also added.
}
\description{
Calculates the bearing angle and distance from each
focal individual of a group to its nearest neighbor over time, across
the sets of a dataset.
}
\seealso{
\code{\link{nn_metrics}, \link{group_metrics_per_set}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
