% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_metrics_calc.R
\name{par_calc_global_metrics}
\alias{par_calc_global_metrics}
\title{Calculation of group metrics in parallel}
\usage{
par_calc_global_metrics(dfs_per_time, lonlat)
}
\arguments{
\item{lonlat}{logical, whether positions are
geographic coordinates, default = FALSE.}

\item{x}{A list of dataframes per timestep.}
}
\value{
A dataframe with the group average timeseries, with columns:
 set, time, pol, speed, shape, N (number of individuals),
 missing_ind (whether some individuals are missing).
}
\description{
Calls calc_global_metrics in parallel
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
