% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_vels.R
\name{parAddVels}
\alias{parAddVels}
\title{Adding motion properties in parallel without progress updates}
\usage{
parAddVels(per_id, lonlat = FALSE)
}
\arguments{
\item{per_id}{Time series with individual's positional data through time.
Columns must include: id, time, lon, lat.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}
}
\value{
the input dataframe with new speed and heading (rotational) column
}
\description{
Calculates headings and speeds based on
two location points and the time taken to travel between those points
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
