% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_motion_timeseries.R
\name{parallel_per_id}
\alias{parallel_per_id}
\title{Speed and heading calculation}
\usage{
parallel_per_id(per_id, lonlat)
}
\arguments{
\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}

\item{data}{Time series of 1 individual's positional data.}
}
\value{
the input dataframe with new speed and heading (rotational) column
}
\description{
Calculates headings and speeds based on two location points of 1 individual
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
