% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perpdot.R
\name{perpDot}
\alias{perpDot}
\title{Perpedicular dot product}
\usage{
perpDot(a, b)
}
\arguments{
\item{a}{a 2d vector}

\item{b}{a 2d vector}
}
\value{
a double that is the perpedicular dot product of a and b
}
\description{
Calculates the perpedicular dot product of 2 vectors
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}, Hanno Hildenbrandt
}
