% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_events.R
\name{pick_events_threshold}
\alias{pick_events_threshold}
\title{Pick thresholds}
\usage{
pick_events_threshold(
  data_distr,
  var,
  interactive = TRUE,
  input_threshold = NA
)
}
\arguments{
\item{data_distr}{A numeric vector to pick a threshold for}

\item{var}{The name of the distribution.}

\item{interactive}{Whether to ask the user for a threshold or not}

\item{input_threshold}{Returned in case interactive is off.}
}
\value{
the input value of the user, representing the
threshold for the definition of an event.
}
\description{
Interactive function that prints the quantiles of the
 input distribution and asks the user to input the threshold value they want.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
