% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_between.R
\name{rad_between}
\alias{rad_between}
\title{Radius between}
\usage{
rad_between(a, b)
}
\arguments{
\item{a}{a 2d vector}

\item{b}{a 2d vector}
}
\value{
a double representing the angle between a and b in rads
}
\description{
Calculates the angle in radius between two vectors.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
