% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data_format.R
\name{set_date_time_format}
\alias{set_date_time_format}
\title{Set format of date and time}
\usage{
set_date_time_format(df_track)
}
\arguments{
\item{df_track}{A trackdf dataframe with a date_time column}
}
\value{
the input dataframe with separated 'date', 'real_time' and 'time' columns
}
\description{
Splits the date_time column of a track dataframe to separate columns and transforms real time to seconds from beginning
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
