% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time2secs.R
\name{transform_time2secs}
\alias{transform_time2secs}
\title{Turns real time to timesteps}
\usage{
transform_time2secs(data, start_t)
}
\arguments{
\item{data}{A dataframe of time series. Column names must contain: time.}

\item{start_t}{The real time from which to start the time series count in seconds.}
}
\value{
The input dataframe with the initial time column as 'real_time' and the overwritten time column in seconds after 'start_t'
}
\description{
Replaces time with seconds since the beginning of the timeseries.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
