% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_mag.R
\name{vector_magnitude}
\alias{vector_magnitude}
\title{Vector magnitude}
\usage{
vector_magnitude(x, y)
}
\arguments{
\item{x}{the x component of the vector}

\item{y}{the y component of the vector}
}
\value{
a double, the magnitude of the input vector
}
\description{
Calculates the magnitude of a 2D vector.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
