% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_pos.R
\name{add_rel_pos_coords}
\alias{add_rel_pos_coords}
\title{Relative position coordinates}
\usage{
add_rel_pos_coords(data, focal_heading = c(0, 1))
}
\arguments{
\item{data}{Dataframe with the bearaing angle and distance of each individual to specific neighbors.
Column names must include: id, time. Can be the output of the neighb_rel_pos_timeseries in long_dataframe format.}

\item{focal_heading}{the heading of the focal individual, default = c(0,1) for plotting neighbor heading north.}
}
\value{
the input dataframe with additional x and y columns
}
\description{
Calculates the x and y coordinates of a neighbor in the reference frame of each focal individual
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
