% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities.R
\name{add_set_vels}
\alias{add_set_vels}
\title{Add Velocity Timeseries}
\usage{
add_set_vels(data, geo = TRUE, verbose = FALSE, parallelize = FALSE)
}
\arguments{
\item{data}{A dataframe with the time series of individuals' positions.
Columns must include: \code{t}, \code{id}, \code{x}, \code{y}.}

\item{geo}{Logical, whether positions are geographic coordinates, default = TRUE.}

\item{verbose}{Logical, whether to post updates on progress, default = FALSE.}

\item{parallelize}{Logical, whether to run the function in parallel, default = FALSE.}
}
\value{
The input dataframe with a new speed and heading (rotational, in rads) columns.
}
\description{
This function calculates the headings and speeds of individuals based on
two location points and the time taken to travel between those points.
}
\seealso{
\code{\link{add_velocities}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
