% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities_ind.R
\name{add_vels}
\alias{add_vels}
\title{Add velocity information}
\usage{
add_vels(data, geo = TRUE, verbose = FALSE, parallelize = FALSE)
}
\arguments{
\item{data}{A dataframe with the time series of individuals' positions.
Columns must include: id, t, x, y.}

\item{geo}{Logical, whether positions are geographic coordinates, default = TRUE.}

\item{verbose}{Logical, whether to post updates on progress, default = FALSE.}

\item{parallelize}{Logical, whether to run the function in parallel, default = FALSE.}
}
\value{
the input dataframe with a new speed and heading (rotational, in rads) columns.
}
\description{
Calculates the headings and speeds of individuals based on
two location points and the time taken to travel between those points.
}
\seealso{
\code{\link{group_vels}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
