% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{calc_dur_per_event}
\alias{calc_dur_per_event}
\title{Events duration}
\usage{
calc_dur_per_event(data, step2time)
}
\arguments{
\item{data}{A dataframe with a keep (representing which rows are defined as events of collective motion) and a set column}

\item{step2time}{the sampling frequency of the dataframe.}
}
\value{
a dataframe of duration per event
}
\description{
Calculates the duration of each event of collective motion in a dataset
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
