% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_global_metrics_analysis.R
\name{calc_global_group_metrics}
\alias{calc_global_group_metrics}
\title{Global group metrics calculation}
\usage{
calc_global_group_metrics(
  data_dates_list,
  mov_av_time_window,
  return_df = TRUE,
  out_csv_dir = NA,
  id_label = "id",
  date_label = "date",
  pos_label_x = "posx",
  pos_label_y = "posy",
  head_label_x = "headx",
  head_label_y = "heady"
)
}
\arguments{
\item{data_dates_list}{A list of dataframes with groups timeseries per day. Column names must include: id, time, date, headx, heady, posx, posy}

\item{mov_av_time_window}{Integer, timesteps to use as a sliding window for average speed and polarization. Set NA if calculation is not needed.}

\item{return_df}{logical, whether to return the result dataframes or not.}

\item{out_csv_dir}{directory output is saved to.}

\item{id_label}{column name of individual local identifier column, default = 'id'.}

\item{date_label}{column name of date, default = 'date'.}

\item{pos_label_x}{column name of x position column, default = 'posx'.}

\item{pos_label_y}{column name of y position column, default = 'posy'.}

\item{head_label_x}{column name of x heading column, default = 'headx'.}

\item{head_label_y}{column name of y heading column, default = 'heady'.}
}
\value{
a dataframe with a column for neighbor id, bearing angle, distance and heading deviation for each individual through time.
}
\description{
Calculates the average speed, polarization and shape of all groups through time.
}
\seealso{
\code{\link{neighb_rel_pos_timeseries_parallel}}, \code{\link{bearing_angle}}, \code{\link{rad_between}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
