% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_global_metrics_analysis.R
\name{calc_global_group_metrics}
\alias{calc_global_group_metrics}
\title{Global group metrics calculation}
\usage{
calc_global_group_metrics(
  data_dates_list,
  mov_av_time_window,
  lonlat,
  return_df = TRUE,
  out_csv_dir = NA
)
}
\arguments{
\item{data_dates_list}{A list of dataframes with groups timeseries per day. Column names must include: id, time, date, headx, heady, posx, posy}

\item{mov_av_time_window}{Integer, timesteps to use as a sliding window for average speed and polarization. Set NA if calculation is not needed.}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}

\item{return_df}{logical, whether to return the result dataframes or not.}

\item{out_csv_dir}{directory output is saved to.}
}
\value{
a dataframe with a column for neighbor id, bearing angle, distance and heading deviation for each individual through time.
}
\description{
Calculates the average speed, polarization and shape of all groups through time.
}
\seealso{
\code{\link{group_metrics_parallel}}, \code{\link{moving_average}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
