% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_global.R
\name{calc_global_metrics}
\alias{calc_global_metrics}
\title{Calculation of group metrics}
\usage{
calc_global_metrics(x, geo)
}
\arguments{
\item{x}{A dataframe of a timestep of a group.}

\item{geo}{Logical, whether positions are
geographic coordinates, default = FALSE.}
}
\value{
A dataframe with one row,
with the group metrics for that timesteps, with columns:
set, t, pol, speed, shape, N (number of individuals),
missing_ind (whether some individuals are missing).
}
\description{
Calculates the average speed,
 polarization and shape of the group through time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
