% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{calc_n_events}
\alias{calc_n_events}
\title{Calculates the number of events}
\usage{
calc_n_events(df)
}
\arguments{
\item{data}{A dataframe with a keep column.}
}
\value{
an integer with the number of events of collective motion
}
\description{
Calculates the number of events
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
