% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{calc_n_events}
\alias{calc_n_events}
\title{Calculates the number of events}
\usage{
calc_n_events(data)
}
\arguments{
\item{data}{The dataframe of one set (splitted by the \code{events_n}
function) with a keep column.}
}
\value{
an integer, the number of events of collective motion
(sequences of keep == TRUE).
}
\description{
Makes the calculation of number of events
of collective motion per set, given by \code{events_n}.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
