% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_metrics_calc.R
\name{calc_nn_metrics}
\alias{calc_nn_metrics}
\title{Relative position of nearest neighbor function to parallelize}
\usage{
calc_nn_metrics(thists, lonlat)
}
\arguments{
\item{thists}{A timestep of individual positions and ids.}

\item{lonlat}{whether positions are geographic coordinates.}
}
\value{
the input dataframe with a column for neighbor id, bearing angle,
distance and heading deviation for each individual through time.
}
\description{
Calculates the bearing angle and distance from a focal
individual of a group to its nearest neighbor.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
