% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_metrics.R
\name{calc_pairwise_metrics}
\alias{calc_pairwise_metrics}
\title{Calculate pairwise metrics}
\usage{
calc_pairwise_metrics(df)
}
\arguments{
\item{df}{data frame with time series of pairwise measurements.
Columns must include:  set, t, id, dist, bangl}
}
\value{
a dataframe with average and sd of pairwise measurement
of a group at a single time step.
}
\description{
Calculating group averages of pairwise measures
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
