% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_col_motion_metrics.R
\name{col_motion_metrics}
\alias{col_motion_metrics}
\title{Collective motion metrics}
\usage{
col_motion_metrics(
  date_dfs,
  global_metrics,
  paiwise_data,
  mov_av_time_window,
  step2time = 1,
  verbose = TRUE,
  interactive_mode = TRUE,
  speed_lim = NA,
  pol_lim = NA
)
}
\arguments{
\item{date_dfs}{A data frame with time series of individual's positional data through time. Columns must include: id, date, time, posx, posy}

\item{global_metrics}{A data frame with the global metrics timeseries.}

\item{paiwise_data}{A data frame with the pairwise metrics timeseries.}

\item{mov_av_time_window}{to average over (in timesteps)}

\item{step2time}{the sampling frequency, the relation between a time step and real time in seconds}

\item{verbose}{whether to post updates on progress}

\item{interactive_mode}{whether to take the threshold for event definition from user.}

\item{speed_lim}{threshold for speed if interactive mode id off}

\item{pol_lim}{threshold for polarization if interactive mode id off}
}
\value{
either a list of dataframes with neighbor ids, bearing angles, distances and heading deviations for each individual through time, or saves individual csvs per day, depending on input.
}
\description{
Calculates metrics of collective motion across dates and events
}
\seealso{
\code{\link{neighb_rel_pos_timeseries_parallel}}, \code{\link{group_motion_timeseries}}, \code{\link{group_metrics_parallel}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
