% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_col_motion_metrics.R
\name{col_motion_metrics}
\alias{col_motion_metrics}
\title{Collective motion metrics}
\usage{
col_motion_metrics(
  timeseries_data,
  global_metrics,
  step2time = 1,
  verbose = TRUE,
  speed_lim = NA,
  pol_lim = NA,
  noise_thresh = 0
)
}
\arguments{
\item{timeseries_data}{A data frame with time series of individual's
positional data through time with nearest neighbor analysis conducted}

\item{global_metrics}{A data frame with the global metrics timeseries.}

\item{step2time}{Numeric, the sampling frequency of the dataset
(the relation between a time step and real time in seconds).}

\item{verbose}{Logical, whether to post updates on progress.}

\item{speed_lim}{Numeric, the threshold of speed for the definition of
an event. For more info see: \code{\link{pick_threshold}}.}

\item{pol_lim}{Numeric, the threshold of polarization for the definition of
an event. For more info see: \code{\link{pick_threshold}}.}

\item{noise_thresh}{Numeric, the limit of time difference between consecutive
events to be considered the same event. Default value is 0 (no event merging).}
}
\value{
A dataframe with metrics of collective motion per event.
}
\description{
Calculates metrics of collective motion across sets
and events.
}
\seealso{
\code{\link{define_events}, \link{group_metrics}, \link{pairwise_metrics}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
