% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_col_motion_metrics.R
\name{col_motion_metrics}
\alias{col_motion_metrics}
\title{Collective motion metrics}
\usage{
col_motion_metrics(
  timeseries_data,
  global_metrics,
  mov_av_time_window,
  step2time = 1,
  verbose = TRUE,
  interactive_mode = TRUE,
  speed_lim = NA,
  pol_lim = NA
)
}
\arguments{
\item{timeseries_data}{A data frame with time series of individual's
positional data through time with nearest neighbor analysis conducted}

\item{global_metrics}{A data frame with the global metrics timeseries.}

\item{mov_av_time_window}{to average over (in timesteps)}

\item{step2time}{the sampling frequency, the relation between a time step
and real time in seconds}

\item{verbose}{whether to post updates on progress}

\item{interactive_mode}{whether to take the threshold for
event definition from user.}

\item{speed_lim}{threshold for speed if interactive mode id off}

\item{pol_lim}{threshold for polarization if interactive mode id off}
}
\value{
a dataframe with metrics of collective motion per event
}
\description{
Calculates metrics of collective motion across sets and events
}
\seealso{
\code{\link{moving_average}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
