% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_col_motion_metrics.R
\name{col_motion_metrics}
\alias{col_motion_metrics}
\title{Collective motion metrics}
\usage{
col_motion_metrics(
  timeseries_data,
  global_metrics,
  step2time = 1,
  verbose = TRUE,
  speed_lim = NA,
  pol_lim = NA,
  noise_thresh = 0
)
}
\arguments{
\item{timeseries_data}{A data frame with time series of individual's
positional data through time with nearest neighbor analysis conducted}

\item{global_metrics}{A data frame with the global metrics timeseries.}

\item{step2time}{the sampling frequency, the relation between a time step
and real time in seconds.}

\item{verbose}{whether to post updates on progress.}

\item{speed_lim}{threshold for speed if interactive mode is off.}

\item{pol_lim}{threshold for polarization if interactive mode is off.}

\item{noise_thresh}{The limit of time difference between consecutive events
to be considered the same event. Default value is 0 (no event merging).}
}
\value{
a dataframe with metrics of collective motion per event
}
\description{
Calculates metrics of collective motion across sets
and events.
}
\seealso{
\code{\link{moving_average}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
