% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_swarm_space.R
\name{create_swarm_space}
\alias{create_swarm_space}
\title{Create swarm space}
\usage{
create_swarm_space(
  metrics_data,
  space_type = "pca",
  event_duration_limit = NA,
  tsne_rand_seed = NA,
  tsne_perplexity = 10
)
}
\arguments{
\item{metrics_data}{Metrics of collective motion per species per event.}

\item{space_type}{Choice between pca and tsne.}

\item{event_duration_limit}{Filter out events that are shorter than given duration.}

\item{tsne_rand_seed}{Random seed for tsne}

\item{tsne_perplexity}{Perplexity parameter for tsne. Usually between 10-50.}
}
\value{
the swarm space, x and y coordinates per event of each species
}
\description{
Runs a pca or a tsne to produce a swarm space.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
