% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities_ind.R
\name{do_add_vels}
\alias{do_add_vels}
\title{Adding velocity information for individuals in parallel without progress updates}
\usage{
do_add_vels(per_id, geo = FALSE)
}
\arguments{
\item{per_id}{Dataframe with the time series of individuals' positional data.
Columns must include: id, t, x, y.}

\item{geo}{Logical, whether positions are geographic coordinates, default = FALSE.}
}
\value{
the input dataframe with a new speed and heading (rotational) column.
}
\description{
Calculates headings and speeds based on
two location points and the time taken to travel between those points.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
