% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_swarm_space.R
\name{do_pca}
\alias{do_pca}
\title{Perform PCA}
\usage{
do_pca(df)
}
\arguments{
\item{df}{a dataframe with the metrics of collective motion
per species/set per event.}
}
\value{
a list with the swarm space (a dataframe with PC1, PC2, PC3
coordinates per event per species) and the new pca object.
}
\description{
Calculating the pca space.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
