% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarm_space.R
\name{do_pca}
\alias{do_pca}
\title{Perform PCA}
\usage{
do_pca(df)
}
\arguments{
\item{df}{A dataframe with the metrics of collective motion
per event.}
}
\value{
a list with the swarm space (a dataframe with PC1, PC2, PC3
coordinates per event) and the complete pca object.
}
\description{
Run a PCA to place events on collective motion into less-dimentional space.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
