% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_idx.R
\name{event_ids}
\alias{event_ids}
\title{Event indexies}
\usage{
event_ids(df)
}
\arguments{
\item{df}{A dataframe with a set and t columns
for adding event ids.}
}
\value{
a vector of the same size as t with event ids
}
\description{
Returns a vector with event ids that corresponds
to the input time vector, according to the continuity of time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
