% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_idx.R
\name{event_ids_per_set}
\alias{event_ids_per_set}
\title{Event indexies per set}
\usage{
event_ids_per_set(setdf, ev_idx)
}
\arguments{
\item{t}{a vector of time.}
}
\value{
a vector of the same size as t with event ids
}
\description{
Returns a vector with event ids that corresponds
to the input time vector, according to the continuity of time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
