% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_idxs.R
\name{event_ids_per_set}
\alias{event_ids_per_set}
\title{Event Indexes per Set}
\usage{
event_ids_per_set(setdf, ev_idx)
}
\arguments{
\item{setdf}{The dataframe for one set of the data. It should not have
discontinuities in time.}

\item{ev_idx}{The last event index added so far at the dataset (e.g., from a
previous set).}
}
\value{
A vector with the timeseries of event indexes for the input set.
}
\description{
The execution of \code{event_ids} per set.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
