% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_N_duration.R
\name{events_dur}
\alias{events_dur}
\title{Total Duration of All Events}
\usage{
events_dur(data, step2time)
}
\arguments{
\item{data}{A dataframe with a \code{keep} column, representing which rows
are defined as events of collective motion (added by the
\code{\link{define_events}} function).}

\item{step2time}{The sampling frequency of the dataframe (how many seconds
are between each row of the data).}
}
\value{
A numeric corresponding to the total duration of events in the
 dataset in seconds.
}
\description{
This function calculates the total duration (in seconds) of
 events of collective motion in a dataset.
}
\seealso{
\code{\link{define_events}}, \code{\link{events_n}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
