% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_global.R
\name{global_metrics}
\alias{global_metrics}
\title{Calculation of group metrics of collective motion}
\usage{
global_metrics(data, geo, step2time = 1, parallelize = FALSE)
}
\arguments{
\item{data}{A dataframe of (ordered) time series of headings,
positions and speeds per individual. The dataframe may contain
several individuals. Should include column for:
id, t, speed, x, y, head, set.}

\item{geo}{Logical, whether positions are geographic coordinates,
default = FALSE.}

\item{step2time}{Numeric, the sampling frequency of the data
(the relation between a row in the data and real time in seconds).}

\item{parallelize}{Logical, whether to parallelize over time.
Suggested only for very large datasets.}
}
\value{
A dataframe with the group average timeseries, with columns:
 set, t, pol, speed, shape, N (number of individuals),
 missing_ind (whether some individuals are missing).
}
\description{
Calculates the average speed, polarization
and shape of the group through time.
}
\seealso{
\code{\link{group_shape}, \link{group_vels}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
