% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_base_metrics.R
\name{group_metrics_parallel}
\alias{group_metrics_parallel}
\title{Calculation of group metrics in parallel}
\usage{
group_metrics_parallel(data, lonlat)
}
\arguments{
\item{data}{A dataframe of (ordered) time series of headings, positions and speeds per individual.
The dataframe may contain several individuals. Should include column for: id, time, speed, headx, heady, posx, posy.}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}
}
\value{
A dataframe with the group average timeseries, with columns:
 set, time, pol, speed, shape, N (number of individuals), missing_ind (whether some individuals are missing).
}
\description{
Calculates the average speed, polarization and shape of the group through time.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
