% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_group_motion_timeseries.R
\name{group_motion_timeseries}
\alias{group_motion_timeseries}
\title{Group collective motion timeseries}
\usage{
group_motion_timeseries(data, lonlat = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{A data frame with time series of individual's positional data through time.}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
a list of dataframes, an element per date from the input dataframe with new columns: headx, heady, velx, vely, speed, real_time
}
\description{
Adds velocity and heading calculations to dataframe and transforms time to seconds from start per day.
}
\seealso{
\code{\link{padd_motion_properties}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
