% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_group_vels.R
\name{group_vels}
\alias{group_vels}
\title{Group velocity timeseries}
\usage{
group_vels(data, lonlat = TRUE, verbose = FALSE, parallelize = FALSE)
}
\arguments{
\item{data}{A data frame with time series of individual's positional data
through time.}

\item{lonlat}{logical, whether positions are geographic coordinates,
default = FALSE.}

\item{verbose}{whether to post updates on progress, default = FALSE.}

\item{parallelize}{whether to run the function in parallel over individuals,
default = FALSE.}
}
\value{
a list of dataframes, an element per set from the input dataframe
with new columns: headx, heady, velx, vely, speed, real_time
}
\description{
Adds velocity and heading calculations to dataframe and
transforms time to seconds from start per day.
}
\seealso{
\code{\link{add_vels}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
