% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Moving average}
\usage{
moving_average(timeseries, window)
}
\arguments{
\item{timeseries}{Vector of doubles representing a timeseries.}

\item{window}{Double, the time-window to average over (in timesteps).}
}
\value{
A vector of doubles (average over the window).
}
\description{
Calculates the moving average of a time series.
}
\examples{
bs <- rnorm(20, mean = 10, sd = 1)
moving_average(bs, 5)

}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
