% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_species_pca_data}
\alias{multi_species_pca_data}
\title{Mutli-species PCA data}
\format{
A dataframe of 3 columns: species, PC1, PC2, PC3.
}
\usage{
data('multi_species_pca_data')
}
\description{
The positions of events from 4 species: stickleback fish, homing pigeons, goats and chacma baboons,
in the PCA swarm space (see *multi_species_pca*). First published as part of:
Papadopoulou Marina, Fürtbauer Ines, O'Bryan Lisa R., Garnier Simon, Georgopoulou Dimitra G., Bracken Anna M., Christensen Charlotte and King Andrew J.
2023. Dynamics of collective motion across time and species. Phil. Trans. R. Soc. B 378: 20220068. http://doi.org/10.1098/rstb.2022.0068
}
\keyword{datasets}
