% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_pairwise.R
\name{nn_metrics}
\alias{nn_metrics}
\title{Nearest neighbour metrics}
\usage{
nn_metrics(
  data,
  add_coords = FALSE,
  geo = FALSE,
  verbose = FALSE,
  parallelize = FALSE
)
}
\arguments{
\item{data}{A dataframe with the group's positional timeseries for one set.
Column names must include: id, x, y, t. The calculations are based on
the \code{swaRm} package.}

\item{add_coords}{Logical, whether the data on relative positions of
nearest neighbours should be converted into
coordinates in the reference frame of the focal individual (nnx, nny). This
can be useful for visualization purposes but it is not used in the package
pipeling. Default = 'FALSE'.}

\item{geo}{Logical, whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{Logical, whether to post updates on progress.}

\item{parallelize}{Logical, whether to parallelize the function over time.}
}
\value{
The input dataframe with new columns for nearest neighbor id (nn_id),
bearing angle (bangl), and distance (nnd).
If add_coords is TRUE, the columns nnx and nny are  added.
}
\description{
Calculates the bearing angle and distance from
all focal individuals in a group to their nearest neighbor over time.
}
\seealso{
\code{\link{add_rel_pos_coords}, \link{group_metrics}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
