% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize Data}
\usage{
normalize_data(vec)
}
\arguments{
\item{vec}{A numerical vector to normalize.}
}
\value{
A vector of doubles, the normalized values of the input vector.
}
\description{
This function rescales a vector to values between 0 and 1.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
