% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_pairwise_analysis.R
\name{pairwise_analysis}
\alias{pairwise_analysis}
\title{Pairwise analysis}
\usage{
pairwise_analysis(
  data_dates_list,
  max_n_rank_to_return = 4,
  date_label = "date",
  return_df = TRUE,
  out_csv_dir = NA,
  out_csv_name = "pairwise_",
  add_coords = FALSE,
  lonlat = FALSE,
  pos_label_x = "lon",
  pos_label_y = "lat",
  head_label_x = "headx",
  head_label_y = "heady",
  verbose = FALSE
)
}
\arguments{
\item{data_dates_list}{A list of dataframes with groups timeseries per day. Column names must include: id, time, date, headx, heady, posx, posy}

\item{max_n_rank_to_return}{Max rank of neighbor (e.g. 4th closest) to save/return data for}

\item{date_label}{column name of date, default = 'date'.}

\item{return_df}{logical, whether to return the result dataframes or not.}

\item{out_csv_dir}{directory output is saved to.}

\item{out_csv_name}{names of files that will be exported.}

\item{add_coords}{whether data is converted to geographic coordinates, default = 'FALSE'.}

\item{lonlat}{logical, whether positions are geographic coordinates, default = FALSE.}

\item{pos_label_x}{column name of x position column, default = 'lon'.}

\item{pos_label_y}{column name of y position column, default = 'lat'.}

\item{head_label_x}{column name of x heading column, default = 'headx'.}

\item{head_label_y}{column name of y heading column, default = 'heady'.}

\item{verbose}{whether to post updates on progress}
}
\value{
either a list of dataframes with neighbor ids, bearing angles, distances and heading deviations for each individual through time, or saves individual csvs per day, depending on input.
}
\description{
Calculates the bearing angle and distance from a focal individual of a group to all neighbors over time.
}
\seealso{
\code{\link{neighb_rel_pos_timeseries_parallel}}, \code{\link{bearing_angle}}, \code{\link{rad_between}}
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
