% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_group.R
\name{par_calc_global_metrics}
\alias{par_calc_global_metrics}
\title{Calculation of Group Metrics in Parallel}
\usage{
par_calc_global_metrics(dfs_per_time, geo)
}
\arguments{
\item{dfs_per_time}{A list of dataframes per timestep.}

\item{geo}{Logical, whether positions are
geographic coordinates, default = FALSE.}
}
\value{
A dataframe with the group average timeseries, with columns:
 \code{set}, \code{t}, \code{pol}, \code{speed}, \code{shape}, \code{N} (number of individuals),
 \code{missing_ind} (whether some individuals are missing).
}
\description{
This function calls calc_global_metrics in parallel.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
