% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_metrics_calc.R
\name{par_nn_metrics}
\alias{par_nn_metrics}
\title{Adding motion properties in parallel - verbose}
\usage{
par_nn_metrics(per_time, lonlat = FALSE)
}
\arguments{
\item{per_time}{Time series with groups positional data through time.
Columns must include: id, time, lon, lat.}

\item{lonlat}{whether positions are geographic coordinates, default = FALSE.}

\item{verbose}{whether to post updates on progress}
}
\value{
the input dataframe with new speed and heading (rotational) column
}
\description{
Calculates headings and speeds based on
two location points and the time taken to travel between those points
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
