% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_calc_pairwise.R
\name{par_nn_metrics}
\alias{par_nn_metrics}
\title{Adding Motion Properties in Parallel - Verbose}
\usage{
par_nn_metrics(per_time, geo = FALSE)
}
\arguments{
\item{per_time}{A groups positional data at a single timestep.
Columns must include: \code{id}, \code{t}, \code{head}, \code{x}, \code{y}.}

\item{geo}{Logical, whether positions are geographic coordinates, default = FALSE.}
}
\value{
The input dataframe with new columns for \code{nnd} (nearest neighbor distance),
 \code{nn_id} (its id) and  \code{bangl} (nearest neighbor bearing angle).
}
\description{
This function calculates the bearing angle and distance from
a focal individual to its nearest neighbor at the given timestep.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
