% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities.R
\name{parAddVels}
\alias{parAddVels}
\title{Adding Motion Properties in Parallel without Progress Updates}
\usage{
parAddVels(per_id, geo = FALSE)
}
\arguments{
\item{per_id}{Dataframe with the time series of individuals' positional data.
Columns must include: \code{t}, \code{id}, \code{x}, \code{y}.}

\item{geo}{Logical, whether positions are geographic coordinates, default = FALSE.}
}
\value{
the input dataframe with a new speed and heading (rotational) column.
}
\description{
This function calculates headings and speeds based on
two location points and the time taken to travel between those points.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
