% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities_ind.R
\name{parAddVels}
\alias{parAddVels}
\title{Adding motion properties in parallel without progress updates}
\usage{
parAddVels(per_id, geo = FALSE)
}
\arguments{
\item{per_id}{Time series of individuals' positional data.
Columns must include: id, t, x, y.}

\item{geo}{Whether positions are geographic coordinates, default = FALSE.}
}
\value{
the input dataframe with a new speed and heading (rotational) column.
}
\description{
Calculates headings and speeds based on
two location points and the time taken to travel between those points.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
