% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocities_ind.R
\name{perIdVels}
\alias{perIdVels}
\title{Speed and heading calculation based on swaRm package.}
\usage{
perIdVels(per_id, geo)
}
\arguments{
\item{geo}{whether positions are geographic coordinates, default = FALSE.}

\item{data}{Time series of 1 individual's positional data.}
}
\value{
the input dataframe with a new speed and heading (rotational) column.
}
\description{
Calculates headings and speeds based on
consecutive locations of 1 individual. Based on the _heading_ and the
_linear\_speed_ functions of _swaRm_.
}
\author{
Marina Papadopoulou \email{m.papadopoulou.rug@gmail.com}
}
\keyword{internal}
